+++ textrun-make/src/actions/target.ts
@@ -1,5 +1,5 @@
-import * as color from "colorette"
 import { promises as fs } from "fs"
+import { styleText } from "node:util"
 import * as textRunner from "text-runner-engine"
 
 import { makefileTargets } from "../helpers/makefile-targets.js"
@@ -10,13 +10,15 @@ export async function target(action: textRunner.actions.Args): Promise<void> {
   if (target === "") {
     throw new Error("Empty make target")
   }
-  action.name(`make target ${color.cyan(target)}`)
+  action.name(`make target ${styleText("cyan", target)}`)
   const makePath = action.configuration.sourceDir.joinStr(action.region[0].attributes["dir"] || ".", "Makefile")
   const makefile = await fs.readFile(makePath, "utf8")
   const targets = makefileTargets(makefile)
   if (!targets.includes(target)) {
     throw new Error(
-      `Makefile does not contain target ${color.cyan(target)} but these ones: ${color.cyan(targets.join(", "))}`
+      `Makefile does not contain target ${styleText("cyan", target)} but these ones: ${
+        styleText("cyan", targets.join(", "))
+      }`
     )
   }
 }