+++ textrun-npm/src/actions/installed-executable.ts
@@ -1,5 +1,5 @@
-import * as color from "colorette"
 import { promises as fsp } from "fs"
+import { styleText } from "node:util"
 import * as path from "path"
 import * as textRunner from "text-runner-engine"
 
@@ -14,13 +14,13 @@ export async function installedExecutable(action: textRunner.actions.Args): Prom
       action.location
     )
   }
-  action.name(`installed npm executable ${color.cyan(commandName)}`)
+  action.name(`installed npm executable ${styleText("cyan", commandName)}`)
   const fullPath = path.join("node_modules", ".bin", commandName)
   try {
     await fsp.stat(action.configuration.sourceDir.joinStr(fullPath))
   } catch (e) {
     if (textRunner.isFsError(e) && e.code === "ENOENT") {
-      throw new Error(`no installed executable ${color.cyan(fullPath)} found`)
+      throw new Error(`no installed executable ${styleText("cyan", fullPath)} found`)
     } else {
       throw e
     }