+++ textrun-workspace/src/index.ts
@@ -1,6 +1,8 @@
 import * as textRunner from "text-runner"
 
-import { additionalFileContent } from "./actions/additional-file-content.js"
+import { appendFile } from "./actions/append-file.js"
+import { compareFiles } from "./actions/compare-files.js"
+import { copyFile } from "./actions/copy-file.js"
 import { emptyFile } from "./actions/empty-file.js"
 import { existingDirectory } from "./actions/existing-directory.js"
 import { existingFileWithContent } from "./actions/existing-file-with-content.js"
@@ -10,7 +12,9 @@ import { newFile } from "./actions/new-file.js"
 import { workingDir } from "./actions/working-dir.js"
 
 export const textrunActions: textRunner.exports.TextrunActions = {
-  additionalFileContent,
+  appendFile,
+  compareFiles,
+  copyFile,
   emptyFile,
   existingDirectory,
   existingFile,